requirements.txt
matplotlib==3.5.1
opencv_python==4.4.0.44
scikit_image==0.16.2
torchvision==0.8.1
pynput==1.7.6
tqdm==4.62.3
pandas==1.2.4
scipy==1.7.1
numpy==1.19.5
torch==1.6.0+cu101
pywin32==227
Pillow==9.1.0
rawpy==0.17.1
scikit_learn==1.0.2
skimage==0.0
tensorboardX==2.5
